Move cms related array functions and tasks to cms/ subdirectories.
