+++ columnflow/tasks/external.py
@@ -6,8 +6,6 @@ Tasks dealing with external data.
 
 from __future__ import annotations
 
-__all__ = []
-
 import os
 import time
 import shutil
@@ -18,10 +16,8 @@ import luigi
 import law
 import order as od
 
-from columnflow.tasks.framework.base import (
-    Requirements, AnalysisTask, ConfigTask, DatasetTask, wrapper_factory,
-)
-from columnflow.util import wget, safe_div
+from columnflow.tasks.framework.base import AnalysisTask, ConfigTask, DatasetTask, wrapper_factory
+from columnflow.util import wget
 
 
 logger = law.logger.get_logger(__name__)
@@ -348,135 +344,3 @@ class BundleExternalFiles(ConfigTask, law.tasks.TransferLocalFile):
 
         # transfer the result
         self.transfer(tmp)
-
-
-class CreatePileupWeights(ConfigTask):
-
-    sandbox = "bash::$CF_BASE/sandboxes/cmssw_default.sh"
-
-    data_mode = luigi.ChoiceParameter(
-        default="hist",
-        choices=["hist", "pileupcalc"],
-        description="the mode to obtain the data pu profile; choices: 'hist', 'pileupcalc'; "
-        "default: 'hist'",
-    )
-    version = None
-
-    # upstream requirements
-    reqs = Requirements(
-        BundleExternalFiles=BundleExternalFiles,
-    )
-
-    def requires(self):
-        return self.reqs.BundleExternalFiles.req(self)
-
-    def output(self):
-        return self.target(f"weights_from_{self.data_mode}.json")
-
-    @law.decorator.log
-    @law.decorator.safe_output
-    def run(self):
-        # prepare the external files and the output weights
-        externals = self.requires()
-        weights = {}
-
-        # read the mc profile
-        mc_profile = self.read_mc_profile_from_cfg(externals.files.pu.mc_profile)
-
-        # loop over nominal and minbias_xs shifts
-        for shift in ["nominal", "minbias_xs_up", "minbias_xs_down"]:
-            # read or create the data profile
-            if self.data_mode == "hist":
-                pu_hist_target = externals.files.pu.data_profile[shift]
-                data_profile = self.read_data_profile_from_hist(pu_hist_target)
-            else:
-                pu_file_target = externals.files.pu.json
-                minbiasxs = self.config_inst.x.minbiasxs.get(shift)
-                data_profile = self.read_data_profile_from_pileupcalc(pu_file_target, minbiasxs)
-
-            # build the ratios and save them
-            if len(mc_profile) != len(data_profile):
-                raise Exception(
-                    f"the number of bins in the MC profile ({len(mc_profile)}) does not match that "
-                    f"of the data profile for the {shift} shift ({len(data_profile)})",
-                )
-
-            # store them
-            weights[shift] = [safe_div(d, m) for m, d in zip(mc_profile, data_profile)]
-            self.publish_message(f"computed pu weights for shift {shift}")
-
-        # save it
-        self.output().dump(weights, formatter="json")
-
-    @classmethod
-    def read_mc_profile_from_cfg(
-        cls,
-        pu_config_target: law.FileSystemTarget,
-    ) -> list[float]:
-        """
-        Takes a mc pileup configuration file stored in *pu_config_target*, parses its content and
-        return the pu profile as a list of float probabilities.
-        """
-        probs = []
-
-        # read non-empty lines
-        lines = map(lambda s: s.strip(), pu_config_target.load(formatter="text").split("\n"))
-
-        # loop through them and extract probabilities
-        found_prob_line = False
-        for line in lines:
-            # look for the start of the pu profile
-            if not found_prob_line:
-                if not line.startswith("mix.input.nbPileupEvents.probValue"):
-                    continue
-                found_prob_line = True
-                line = line.split("(", 1)[-1].strip()
-            # skip empty lines
-            if not line:
-                continue
-            # extract numbers
-            probs.extend(map(float, filter(bool, line.split(")", 1)[0].split(","))))
-            # look for the end of the pu profile
-            if ")" in line:
-                break
-
-        return cls.normalize_values(probs)
-
-    @classmethod
-    def read_data_profile_from_hist(
-        cls,
-        pu_hist_target: law.FileSystemTarget,
-    ) -> list[float]:
-        """
-        Takes the pileup profile in data preproducd by the lumi pog and stored in *pu_hist_target*,
-        builds the ratio to mc and returns the weights in a list.
-        """
-        hist_file = pu_hist_target.load(formatter="uproot")
-        probs = hist_file["pileup"].values().tolist()
-        return cls.normalize_values(probs)
-
-    @classmethod
-    def read_data_profile_from_pileupcalc(
-        cls,
-        pu_file_target: law.FileSystemTarget,
-        minbias_xs: float,
-    ) -> list[float]:
-        """
-        Takes the pileup profile in data read stored in *pu_file_target*, which should have been
-        produced when processing data, and a *minbias_mexs* value in mb (milli), builds the ratio to
-        mc and returns the weights in a list for 99 bins (recommended number).
-        """
-        raise NotImplementedError
-
-    @classmethod
-    def normalize_values(cls, values: Sequence[float]) -> list[float]:
-        _sum = sum(values, 0.0)
-        return [value / _sum for value in values]
-
-
-CreatePileupWeightsWrapper = wrapper_factory(
-    base_cls=AnalysisTask,
-    require_cls=CreatePileupWeights,
-    enable=["configs", "skip_configs"],
-    attributes={"version": None},
-)
