+++ law.cfg
@@ -10,7 +10,6 @@ columnflow.tasks.ml
 columnflow.tasks.union
 columnflow.tasks.histograms
 columnflow.tasks.plotting
-columnflow.tasks.inference
 columnflow.tasks.cutflow
